Forgot the typecast

+++ Engine/src/engine.c
@@ -1341,7 +1341,7 @@ static void grouscan (int32_t dax1, int32_t dax2, int32_t sectnum, uint8_t  dast
                       g_x1, g_y1,
                       a3, // asm3
                       g_x3, g_y3,
-                      (float)-(zd >> (16-BITSOFPRECISION)), //asm1
+                      -(zd >> (16-BITSOFPRECISION)), //asm1
                       -ylookup[1],
                       &tiles[picnum]);
